# hhspec/specs/architecture/api-contracts.yml
...
openapi: "3.0.3"
info:
  title: "WellAlly Health Web API"
  version: "1.0.0"
  description: "个人医疗健康数据管理 Web 界面后端 API"
servers:
  - url: "http://localhost:3000/api"
    description: "本地开发服务器"
security:
  - bearerAuth: []
paths:
...
  /reports/upload:
    post:
      operationId: uploadReport
      summary: 上传检查单图片并触发 AI 识别
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image, report_date, report_type]
              properties:
                image:
                  type: string
                  format: binary
                  description: 检查单图片 (jpg/png/pdf, ≤20MB)
                report_date:
                  type: string
                  format: date
                report_type:
                  type: string
                  enum: [biochemical, imaging]
                hospital_name:
                  type: string
                  maxLength: 300
                department:
                  type: string
                  maxLength: 200
      responses:
        201:
          description: 上传成功，AI 识别中
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthReport"
        400:
          $ref: "#/components/responses/BadRequest"
        413:
          description: 文件过大
  /reports:
    get:
      operationId: listReports
      summary: 查询健康报告列表
      security: [{ bearerAuth: [] }]
      parameters:
        - name: report_type
          in: query
          schema: { type: string, enum: [biochemical, imaging] }
        - name: start_date
          in: query
          schema: { type: string, format: date }
        - name: end_date
          in: query
          schema: { type: string, format: date }
        - name: status
          in: query
          schema: { type: string, enum: [uploaded, ai_processing, processed, review_needed, confirmed, failed] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        200:
          description: 报告列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/HealthReport"
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
  /reports/{id}:
    get:
      operationId: getReport
      summary: 获取报告详情（含指标/发现）
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        200:
          description: 报告详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthReportDetail"
        404:
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateReport
      summary: 更新报告（人工复核修正）
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                indicators:
                  type: array
                  items:
                    $ref: "#/components/schemas/IndicatorUpdate"
      responses:
        200:
          description: 更新成功
    delete:
      operationId: deleteReport
      summary: 删除报告
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        204:
          description: 删除成功

  # ==================== Medications ====================
  /medications:
    get:
      operationId: listMedications
      summary: 查询用药列表
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, discontinued, completed, paused] }
      responses:
        200:
          description: 用药列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: addMedication
      summary: 添加用药记录
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [drug_name, dosage, frequency, start_date]
              properties:
                drug_name: { type: string, maxLength: 300 }
                generic_name: { type: string, maxLength: 300 }
                dosage: { type: string, maxLength: 100 }
                dosage_unit: { type: string, maxLength: 50 }
                frequency: { type: string, maxLength: 100 }
                route: { type: string, enum: [oral, topical, injection, inhalation, sublingual, rectal, other] }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
                notes: { type: string }
      responses:
        201:
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
...
  /consultations:
    get:
      operationId: listConsultations
      summary: 查询咨询历史
      security: [{ bearerAuth: [] }]
      responses:
        200:
          description: 咨询列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"

  # ==================== Trends ====================
...
  /cga-assessments:
    get:
      operationId: listCGAAssessments
      summary: 查询 CGA 评估列表
      security: [{ bearerAuth: [] }]
      responses:
        200:
          description: CGA 评估列表
    post:
      operationId: createCGAAssessment
      summary: 创建新的 CGA 评估
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assessment_date]
              properties:
                assessment_date: { type: string, format: date }
      responses:
        201:
          description: 创建成功
...
  /cognitive-screenings:
    get:
      operationId: listScreenings
      summary: 查询认知筛查列表
      security: [{ bearerAuth: [] }]
      responses:
        200:
          description: 筛查列表
    post:
      operationId: createScreening
      summary: 创建新的认知筛查
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [screening_date, screening_type]
              properties:
                screening_date: { type: string, format: date }
                screening_type: { type: string, enum: [AD8, MMSE, MoCA, CDR, Mini_Cog, GPCOG] }
                education_years: { type: integer }
                informant_available: { type: boolean }
                cga_assessment_id: { type: integer }
...
  /profile:
    get:
      operationId: getProfile
      summary: 获取个人档案
      security: [{ bearerAuth: [] }]
      responses:
        200:
          description: 个人档案
    put:
      operationId: updateProfile
      summary: 更新个人档案
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string }
                height_cm: { type: number }
                weight_kg: { type: number }
                birth_date: { type: string, format: date }
                gender: { type: string, enum: [male, female, other] }
                blood_type: { type: string, enum: [A, B, AB, O] }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    HealthReport:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        report_type: { type: string }
        title: { type: string }
        report_date: { type: string, format: date }
        hospital_name: { type: string, nullable: true }
        department: { type: string, nullable: true }
        ai_processed: { type: boolean }
        status: { type: string }
        created_at: { type: string, format: date-time }
    HealthReportDetail:
      allOf:
        - $ref: "#/components/schemas/HealthReport"
        - type: object
          properties:
            indicators:
              type: array
              items:
                $ref: "#/components/schemas/BiochemicalIndicator"
              description: 仅生化报告
            findings:
              type: array
              items:
                $ref: "#/components/schemas/ImagingFinding"
              description: 仅影像报告
    BiochemicalIndicator:
      type: object
      properties:
        id: { type: integer }
        indicator_name: { type: string }
        indicator_code: { type: string }
        value: { type: number }
        unit: { type: string }
        reference_range_low: { type: number, nullable: true }
        reference_range_high: { type: number, nullable: true }
        reference_range_text: { type: string, nullable: true }
        is_abnormal: { type: boolean }
        abnormality_direction: { type: string, nullable: true }
    ImagingFinding:
      type: object
      properties:
        id: { type: integer }
        body_part: { type: string }
        modality: { type: string }
        finding: { type: string }
        impression: { type: string, nullable: true }
    IndicatorUpdate:
      type: object
      properties:
        id: { type: integer }
        value: { type: number }
        indicator_name: { type: string }
    Medication:
      type: object
      properties:
        id: { type: integer }
        drug_name: { type: string }
        generic_name: { type: string, nullable: true }
        dosage: { type: string }
        dosage_unit: { type: string }
        frequency: { type: string }
        route: { type: string }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date, nullable: true }
        status: { type: string }
        notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
    Consultation:
      type: object
      properties:
        id: { type: integer }
        consultation_type: { type: string }
        specialty: { type: string, nullable: true }
        chief_complaint: { type: string }
        ai_response: { type: string, nullable: true }
        mdt_specialties: { type: array, items: { type: string }, nullable: true }
        mdt_summary: { type: string, nullable: true }
        status: { type: string }
        created_at: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...